[CHORE] 리뷰 등록 과정 수정 (#84)
