+++ FoodBowl/Global/Base/CategoryCollectionViewCell.swift
