+++ FoodBowl/Screens/DeplicatedMain/UIComponent/StoreInfoView.swift
