--- FoodBowl/Screens/Feed/SetCommentViewController.swift
@@ -1,141 +1,0 @@
-//
-//  SetCommentViewController.swift
-//  FoodBowl
-//
-//  Created by COBY_PRO on 2022/12/23.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class SetCommentViewController: BaseViewController {
-    var delegate: SetCommentViewControllerDelegate?
-
-    private enum Size {
-        static let cellWidth: CGFloat = 100
-        static let cellHeight: CGFloat = cellWidth
-        static let collectionInset = UIEdgeInsets(
-            top: 0,
-            left: 20,
-            bottom: 0,
-            right: 20
-        )
-    }
-
-    var selectedImages = [UIImage]()
-
-    private let textViewPlaceHolder = "ex. "
-
-    // MARK: - property
-
-    private let guideLabel = UILabel().then {
-        $0.text = "후기를 작성해주세요."
-        $0.font = UIFont.preferredFont(forTextStyle: .title3, weight: .medium)
-        $0.textColor = .mainText
-    }
-
-    private let collectionViewFlowLayout = UICollectionViewFlowLayout().then {
-        $0.scrollDirection = .horizontal
-        $0.sectionInset = Size.collectionInset
-        $0.itemSize = CGSize(width: Size.cellWidth, height: Size.cellHeight)
-        $0.minimumInteritemSpacing = 4
-    }
-
-    private lazy var listCollectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewFlowLayout).then {
-        $0.dataSource = self
-        $0.delegate = self
-        $0.showsHorizontalScrollIndicator = false
-        $0.register(FeedThumnailCollectionViewCell.self, forCellWithReuseIdentifier: FeedThumnailCollectionViewCell.className)
-        $0.backgroundColor = .clear
-    }
-
-    lazy var commentTextView = UITextView().then {
-        $0.textContainerInset = UIEdgeInsets(top: 12, left: 10, bottom: 12, right: 10)
-        $0.font = UIFont.preferredFont(forTextStyle: .subheadline, weight: .light)
-        $0.textAlignment = NSTextAlignment.left
-        $0.dataDetectorTypes = UIDataDetectorTypes.all
-        $0.text = textViewPlaceHolder
-        $0.textColor = .grey001
-        $0.isEditable = true
-        $0.delegate = self
-        $0.isScrollEnabled = true
-        $0.isUserInteractionEnabled = true
-        $0.makeBorderLayer(color: .grey002)
-        $0.backgroundColor = .clear
-    }
-
-    // MARK: - life cycle
-
-    override func setupLayout() {
-        view.addSubviews(guideLabel, listCollectionView, commentTextView)
-
-        guideLabel.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide.snp.top).inset(20)
-            $0.leading.equalToSuperview().inset(20)
-        }
-
-        listCollectionView.snp.makeConstraints {
-            $0.top.equalTo(guideLabel.snp.bottom).offset(20)
-            $0.leading.trailing.equalToSuperview()
-            $0.height.equalTo(100)
-        }
-
-        commentTextView.snp.makeConstraints {
-            $0.top.equalTo(listCollectionView.snp.bottom).offset(10)
-            $0.leading.trailing.equalToSuperview().inset(20)
-            $0.bottom.equalTo(view.keyboardLayoutGuide.snp.top).offset(-40)
-        }
-    }
-}
-
-extension SetCommentViewController: UITextViewDelegate {
-    func textViewDidBeginEditing(_ textView: UITextView) {
-        if textView.text == textViewPlaceHolder {
-            textView.text = nil
-            textView.textColor = .mainText
-        }
-    }
-
-    func textViewDidEndEditing(_ textView: UITextView) {
-        if textView.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            textView.text = textViewPlaceHolder
-            textView.textColor = .grey001
-        } else {
-            delegate?.setComment(comment: textView.text)
-        }
-    }
-}
-
-extension SetCommentViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-    func collectionView(_: UICollectionView, numberOfItemsInSection _: Int) -> Int {
-        return selectedImages.count
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(
-            withReuseIdentifier: FeedThumnailCollectionViewCell.className,
-            for: indexPath
-        ) as? FeedThumnailCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-
-        cell.thumnailImageView.image = selectedImages[indexPath.item]
-
-        return cell
-    }
-}
-
-protocol SetCommentViewControllerDelegate: AnyObject {
-    func setComment(comment: String?)
-}
-
-extension SetCommentViewController: SetPhotoViewControllerDelegateForComment {
-    func setPhotoes(photoes: [UIImage]?) {
-        if let photoes = photoes {
-            selectedImages = photoes
-            listCollectionView.reloadData()
-        }
-    }
-}
